Splitted Where clause from Query class.
