+++ src/Select.php
@@ -2,6 +2,7 @@
 
 namespace Girgias\QueryBuilder;
 
+use Girgias\QueryBuilder\Clauses\Where;
 use Girgias\QueryBuilder\Enums\AggregateFunctions;
 use Girgias\QueryBuilder\Enums\SqlOperators;
 use Girgias\QueryBuilder\Exceptions\DangerousSqlQueryWarning;
@@ -15,6 +16,8 @@ use RuntimeException;
 
 class Select extends Query
 {
+    use Where;
+
     /**
      * @var ?string
      */
